Improved logging and adding proper node handling

+++ db/sql/plan/execution_plan.go
@@ -2,5 +2,5 @@ package plan
 
 type NodeExecutionPlan struct {
 	CompiledQuery string
-	NodeID        int
+	NodeID        uint64
 }
